Implemented runNCycles
FPGA error reporting and now just need to check all this and then implement capturing at a lower rate for runNCycles which atm supports up to 255 cycles with graphical feedback in the ui after.

+++ Code/C++/Shared/ADC/MonitorVoltageThresholdHandle.hpp
@@ -13,7 +13,7 @@ private:
     // Thread-safe current voltage (atomic)
     std::atomic<uint16_t> _currentVoltageRaw;
     // Raw ADC threshold (12-bit). Thread-safe.
-    std::atomic<uint32_t> _rawThreshold;
+    std::atomic<uint16_t> _rawThreshold;
 
 public:
